requirements/base.txt
attrs==17.4.0
base58==0.2.5
cffi==1.11.4
coverage==4.5.1
defaultcontext==1.0.3
docutils==0.14
funcsigs==1.0.2
future==0.16.0
-e git+https://github.com/gdanezis/rousseau-chain@eb0bf275eba288e9af1cbbbaa3ef8e63e6b63d0b#egg=hippiehug&subdirectory=hippiehug-package
msgpack-python==0.5.4
petlib==0.0.43
pluggy==0.6.0
profiled==0.1.0
py==1.5.2
pycparser==2.18
pytest==3.4.1
pytest-cov==2.5.1
PyYAML==3.12
redis==2.10.6
six==1.11.0
statistics==1.0.3.5
